experiments/requirements.txt
openai
python-dotenv
google-genai
langchain_core
langchain_text_splitters